Move bootstrap to after "Configure Applications"
Rename the Deploy screen to Configure Applications and move Bootstrap to
after it.  Rename Deploy Status to Deploy and make it actually perform
the deploy.  Combine Bootstrap and BootstrapWait to do both.
