+++ conjureup/controllers/deploy/common.py
@@ -1,20 +1,47 @@
-import os
-from datetime import datetime
+import asyncio
+from operator import attrgetter
 
-from conjureup import events, utils
+from conjureup import events, juju, utils
 from conjureup.app_config import app
-from conjureup.bundlewriter import BundleWriter
 from conjureup.models.step import StepModel
 
 
-def write_bundle(assignments):
-    bundle = app.metadata_controller.bundle
-    bw = BundleWriter(assignments, bundle)
-    datetimestr = datetime.now().strftime("%Y-%m-%d-%H:%M:%S")
-    fn = os.path.join(app.env['CONJURE_UP_CACHEDIR'],
-                      '{}-deployed-{}.yaml'.format(app.env['CONJURE_UP_SPELL'],
-                                                   datetimestr))
-    bw.write_bundle(fn)
+async def do_deploy(msg_cb):
+    await events.ModelConnected.wait()
+    cloud_types = juju.get_cloud_types_by_name()
+    default_series = app.metadata_controller.series
+    machines = app.metadata_controller.bundle.machines
+    applications = sorted(app.metadata_controller.bundle.services,
+                          key=attrgetter('service_name'))
+
+    await pre_deploy(msg_cb=msg_cb)
+    machine_map = await juju.add_machines(applications,
+                                          machines,
+                                          msg_cb=msg_cb)
+    tasks = []
+    for service in applications:
+        if cloud_types[app.current_cloud] == "localhost":
+            # ignore placement when deploying to localhost
+            service.placement_spec = None
+        elif service.placement_spec:
+            # remap machine references to actual deployed machine IDs
+            # (they will only ever not already match if deploying to
+            # an existing model that has other machines)
+            new_placements = []
+            for plabel in service.placement_spec:
+                if ':' in plabel:
+                    ptype, pid = plabel.split(':')
+                    new_placements.append(':'.join([ptype, machine_map[pid]]))
+                else:
+                    new_placements.append(machine_map[plabel])
+            service.placement_spec = new_placements
+
+        tasks.append(juju.deploy_service(service, default_series,
+                                         msg_cb=msg_cb))
+        tasks.append(juju.set_relations(service,
+                                        msg_cb=msg_cb))
+    await asyncio.gather(*tasks)
+    events.DeploymentComplete.set()
 
 
 async def pre_deploy(msg_cb):
@@ -34,3 +61,20 @@ async def pre_deploy(msg_cb):
     await utils.run_step(step,
                          msg_cb)
     events.PreDeployComplete.set()
+
+
+async def wait_for_applications(msg_cb):
+    await events.DeploymentComplete.wait()
+    msg = 'Waiting for deployment to settle.'
+    app.log.info(msg)
+    msg_cb(msg)
+
+    step = StepModel({'title': 'Deployment Watcher'},
+                     filename='00_deploy-done',
+                     name='00_deploy-done')
+    await utils.run_step(step, msg_cb)
+
+    events.ModelSettled.set()
+    msg = 'Model settled.'
+    app.log.info(msg)
+    msg_cb(msg)
