+++ conjureup/controllers/deploy/tui.py
@@ -1,7 +1,4 @@
-import asyncio
-from operator import attrgetter
-
-from conjureup import controllers, events, juju, utils
+from conjureup import controllers, utils
 from conjureup.app_config import app
 
 from . import common
@@ -9,31 +6,12 @@ from . import common
 
 class DeployController:
     def render(self):
-        app.loop.create_task(self.do_deploy())
+        app.loop.create_task(common.do_deploy(utils.info))
+        app.loop.create_task(self._wait_for_applications())
 
-    async def do_deploy(self):
-        cloud_types = juju.get_cloud_types_by_name()
-        default_series = app.metadata_controller.series
-        machines = app.metadata_controller.bundle.machines
-        applications = sorted(app.metadata_controller.bundle.services,
-                              key=attrgetter('service_name'))
-
-        await common.pre_deploy(msg_cb=utils.info)
-        await juju.add_machines(applications,
-                                machines,
-                                msg_cb=utils.info)
-        tasks = []
-        for service in applications:
-            # ignore placement when deploying to localhost
-            if cloud_types[app.current_cloud] == "localhost":
-                service.placement_spec = None
-            tasks.append(juju.deploy_service(service, default_series,
-                                             msg_cb=utils.info))
-            tasks.append(juju.set_relations(service,
-                                            msg_cb=utils.info))
-        await asyncio.gather(*tasks)
-        events.DeploymentComplete.set()
-        controllers.use('deploystatus').render()
+    async def _wait_for_applications(self):
+        await common.wait_for_applications(utils.info)
+        return controllers.use('runsteps').render()
 
 
 _controller_class = DeployController
