+++ conjureup/controllers/showsteps/gui.py
@@ -37,7 +37,7 @@ class ShowStepsController:
             app.steps_data[step.name][field.key] = field.input.value
 
     def finish(self):
-        return controllers.use('bootstrap').render()
+        return controllers.use('configapps').render()
 
 
 _controller_class = ShowStepsController
