+++ conjureup/controllers/showsteps/tui.py
@@ -5,7 +5,7 @@ from conjureup.controllers.showsteps import common
 class ShowStepsController:
     def render(self):
         common.load_steps()
-        controllers.use('bootstrap').render()
+        controllers.use('configapps').render()
 
 
 _controller_class = ShowStepsController
